Projekat2/streaming/requirements.txt
pyspark
python-dotenv
cassandra-driver
